rewrite shell scripts to their docker counterpart
hugo has changed a lot and the only way the current template would work
is to run it with an older version of hugo. the easiest way would be to
run hugo via docker, but that requires rewriting the scripts that expect
hugo to be installed locally to run inside docker instead.